requirements.txt
pandas==2.2.2
openpyxl==3.1.5
matplotlib==3.9.0
pyyaml==6.0.2